+++ TestManagement/src/main/java/com/cg/TestManagement/service/Service.java
@@ -1,6 +1,5 @@
 package com.cg.TestManagement.service;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 
@@ -12,22 +11,22 @@ import com.cg.TestManagement.exception.UserException;
 
 public interface Service {
 	public OnlineTest addTest(OnlineTest onlineTest) throws UserException;
-	public OnlineTest updateTest(BigInteger testId, OnlineTest onlineTest) throws UserException;
-	public OnlineTest deleteTest(BigInteger testId) throws UserException;
-	public Question addQuestion(BigInteger testId, Question question) throws UserException;
-	public Question updateQuestion(BigInteger testId, BigInteger questionId, Question question) throws UserException;
-	public Question deleteQuestion(BigInteger testId, BigInteger questionId) throws UserException;
+	public OnlineTest updateTest(Long testId, OnlineTest onlineTest) throws UserException;
+	public OnlineTest deleteTest(Long testId) throws UserException;
+	public Question addQuestion(Long testId, Question question) throws UserException;
+	public Question updateQuestion(Long testId, Long questionId, Question question) throws UserException;
+	public Question deleteQuestion(Long testId, Long questionId) throws UserException;
 	public User registerUser(User user) throws UserException;
 	public Boolean answerQuestion(OnlineTest onlineTest, Question question, Integer chosenAnswer) throws UserException;
-	public Question showQuestion(OnlineTest onlineTest, BigInteger questionId) throws UserException;
-	public Boolean assignTest(BigInteger userId, BigInteger testId) throws UserException;
+	public Question showQuestion(OnlineTest onlineTest, Long questionId) throws UserException;
+	public Boolean assignTest(Long userId, Long testId) throws UserException;
 	public Double getResult(OnlineTest onlineTest) throws UserException;
 	public Double calculateTotalMarks(OnlineTest onlineTest) throws UserException;
-	public OnlineTest searchTest(BigInteger testId) throws UserException;
-	public User searchUser(BigInteger userId) throws UserException;
-	public void validateUserId(BigInteger id) throws UserException;
-	public void validateTestId(BigInteger id) throws UserException;
-	public void validateQuestionId(BigInteger id) throws UserException;
+	public OnlineTest searchTest(Long testId) throws UserException;
+	public User searchUser(Long userId) throws UserException;
+	public void validateUserId(Long id) throws UserException;
+	public void validateTestId(Long id) throws UserException;
+	public void validateQuestionId(Long id) throws UserException;
 	public void validateUserName(String name) throws UserException;
 	public void validatePassword(String password) throws UserException;
 	public void validateDate(LocalDateTime startDate, LocalDateTime endDate) throws UserException;
